ai_chat/requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
google-genai==1.2.0

# Speech-to-Text
google-cloud-speech==2.28.0
pyaudio==0.2.14

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.27.0
